@@ -1,19 +1,18 @@
 name: photo_view
-description: Photo View provides a gesture sensitive zoomable widget. Photo View is largely used to show interacive images and other stuff such as SVG.
-version: 0.9.0
-author: Renan C. Araújo <renan@caraujo.me>
+description: Photo View provides a gesture sensitive zoomable widget. Photo View is largely used to show interactive images and other stuff such as SVG.
+version: 0.11.1
 homepage: https://github.com/renancaraujo/photo_view
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
-  flutter: ">=1.5.9-pre.94 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.6.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  test: ^1.5.1
+  test: ^1.16.0-nullsafety.17
   flutter_test:
     sdk: flutter
 
