@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:photo_view/photo_view.dart';
 import 'package:photo_view/photo_view_gallery.dart';
-import 'package:photo_view_example/screens/app_bar.dart';
+import 'package:photo_view_example/screens/common/app_bar.dart';
 import 'package:photo_view_example/screens/examples/gallery/gallery_example_item.dart';
 
 class GalleryExample extends StatefulWidget {
@@ -15,61 +15,52 @@ class _GalleryExampleState extends State<GalleryExample> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: <Widget>[
-          const ExampleAppBar(
-            title: "Gallery Example",
-            showGoBack: true,
-          ),
-          Expanded(
-            child: Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      GalleryExampleItemThumbnail(
-                        galleryExampleItem: galleryItems[0],
-                        onTap: () {
-                          open(context, 0);
-                        },
-                      ),
-                      GalleryExampleItemThumbnail(
-                        galleryExampleItem: galleryItems[2],
-                        onTap: () {
-                          open(context, 2);
-                        },
-                      ),
-                      GalleryExampleItemThumbnail(
-                        galleryExampleItem: galleryItems[3],
-                        onTap: () {
-                          open(context, 3);
-                        },
-                      ),
-                    ],
-                  ),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      const Text("Vertical"),
-                      Checkbox(
-                        value: verticalGallery,
-                        onChanged: (value) {
-                          setState(() {
-                            verticalGallery = value;
-                          });
-                        },
-                      ),
-                    ],
-                  ),
-                ],
-              ),
+    return ExampleAppBarLayout(
+      title: "Gallery Example",
+      showGoBack: true,
+      child: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                GalleryExampleItemThumbnail(
+                  galleryExampleItem: galleryItems[0],
+                  onTap: () {
+                    open(context, 0);
+                  },
+                ),
+                GalleryExampleItemThumbnail(
+                  galleryExampleItem: galleryItems[2],
+                  onTap: () {
+                    open(context, 2);
+                  },
+                ),
+                GalleryExampleItemThumbnail(
+                  galleryExampleItem: galleryItems[3],
+                  onTap: () {
+                    open(context, 3);
+                  },
+                ),
+              ],
             ),
-          ),
-        ],
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                const Text("Vertical"),
+                Checkbox(
+                  value: verticalGallery,
+                  onChanged: (value) {
+                    setState(() {
+                      verticalGallery = value!;
+                    });
+                  },
+                ),
+              ],
+            ),
+          ],
+        ),
       ),
     );
   }
@@ -93,17 +84,17 @@ class _GalleryExampleState extends State<GalleryExample> {
 
 class GalleryPhotoViewWrapper extends StatefulWidget {
   GalleryPhotoViewWrapper({
-    this.loadingChild,
+    this.loadingBuilder,
     this.backgroundDecoration,
     this.minScale,
     this.maxScale,
-    this.initialIndex,
-    @required this.galleryItems,
+    this.initialIndex = 0,
+    required this.galleryItems,
     this.scrollDirection = Axis.horizontal,
   }) : pageController = PageController(initialPage: initialIndex);
 
-  final Widget loadingChild;
-  final Decoration backgroundDecoration;
+  final LoadingBuilder? loadingBuilder;
+  final BoxDecoration? backgroundDecoration;
   final dynamic minScale;
   final dynamic maxScale;
   final int initialIndex;
@@ -118,13 +109,7 @@ class GalleryPhotoViewWrapper extends StatefulWidget {
 }
 
 class _GalleryPhotoViewWrapperState extends State<GalleryPhotoViewWrapper> {
-  int currentIndex;
-
-  @override
-  void initState() {
-    currentIndex = widget.initialIndex;
-    super.initState();
-  }
+  late int currentIndex = widget.initialIndex;
 
   void onPageChanged(int index) {
     setState(() {
@@ -146,8 +131,8 @@ class _GalleryPhotoViewWrapperState extends State<GalleryPhotoViewWrapper> {
             PhotoViewGallery.builder(
               scrollPhysics: const BouncingScrollPhysics(),
               builder: _buildItem,
-              itemCount: galleryItems.length,
-              loadingChild: widget.loadingChild,
+              itemCount: widget.galleryItems.length,
+              loadingBuilder: widget.loadingBuilder,
               backgroundDecoration: widget.backgroundDecoration,
               pageController: widget.pageController,
               onPageChanged: onPageChanged,
@@ -185,14 +170,14 @@ class _GalleryPhotoViewWrapperState extends State<GalleryPhotoViewWrapper> {
             childSize: const Size(300, 300),
             initialScale: PhotoViewComputedScale.contained,
             minScale: PhotoViewComputedScale.contained * (0.5 + index / 10),
-            maxScale: PhotoViewComputedScale.covered * 1.1,
+            maxScale: PhotoViewComputedScale.covered * 4.1,
             heroAttributes: PhotoViewHeroAttributes(tag: item.id),
           )
         : PhotoViewGalleryPageOptions(
             imageProvider: AssetImage(item.resource),
             initialScale: PhotoViewComputedScale.contained,
             minScale: PhotoViewComputedScale.contained * (0.5 + index / 10),
-            maxScale: PhotoViewComputedScale.covered * 1.1,
+            maxScale: PhotoViewComputedScale.covered * 4.1,
             heroAttributes: PhotoViewHeroAttributes(tag: item.id),
           );
   }
