@@ -1,58 +1,54 @@
 import 'package:flutter/material.dart';
 import 'package:photo_view/photo_view.dart';
-import 'package:photo_view_example/screens/app_bar.dart';
+import 'package:photo_view_example/screens/common/app_bar.dart';
 
 class HeroExample extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: <Widget>[
-          const ExampleAppBar(
-            title: "Hero Example",
-            showGoBack: true,
-          ),
-          Expanded(
-            child: Center(
-              child: GestureDetector(
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const HeroPhotoViewWrapper(
-                        imageProvider: AssetImage("assets/large-image.jpg"),
-                      ),
-                    ),
-                  );
-                },
-                child: Container(
-                  child: Hero(
-                    tag: "someTag",
-                    child: Image.asset("assets/large-image.jpg", width: 150.0),
+    return ExampleAppBarLayout(
+      title: "Hero",
+      showGoBack: true,
+      child: Center(
+        child: GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => const HeroPhotoViewRouteWrapper(
+                  imageProvider: NetworkImage(
+                    "https://source.unsplash.com/4900x3600/?camera,paper",
                   ),
                 ),
               ),
+            );
+          },
+          child: Container(
+            child: Hero(
+              tag: "someTag",
+              child: Image.network(
+                "https://source.unsplash.com/4900x3600/?camera,paper",
+                width: 350.0,
+                loadingBuilder: (_, child, chunk) =>
+                    chunk != null ? const Text("loading") : child,
+              ),
             ),
-          )
-        ],
+          ),
+        ),
       ),
     );
   }
 }
 
-class HeroPhotoViewWrapper extends StatelessWidget {
-  const HeroPhotoViewWrapper({
-    this.imageProvider,
-    this.loadingChild,
+class HeroPhotoViewRouteWrapper extends StatelessWidget {
+  const HeroPhotoViewRouteWrapper({
+    required this.imageProvider,
     this.backgroundDecoration,
     this.minScale,
     this.maxScale,
   });
 
   final ImageProvider imageProvider;
-  final Widget loadingChild;
-  final Decoration backgroundDecoration;
+  final BoxDecoration? backgroundDecoration;
   final dynamic minScale;
   final dynamic maxScale;
 
@@ -64,7 +60,6 @@ class HeroPhotoViewWrapper extends StatelessWidget {
       ),
       child: PhotoView(
         imageProvider: imageProvider,
-        loadingChild: loadingChild,
         backgroundDecoration: backgroundDecoration,
         minScale: minScale,
         maxScale: maxScale,
