@@ -1,8 +1,10 @@
 library photo_view_gallery;
 
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:photo_view/photo_view.dart'
     show
+        LoadingBuilder,
         PhotoView,
         PhotoViewImageTapDownCallback,
         PhotoViewImageTapUpCallback,
@@ -23,7 +25,7 @@ typedef PhotoViewGalleryBuilder = PhotoViewGalleryPageOptions Function(
 
 /// A [StatefulWidget] that shows multiple [PhotoView] widgets in a [PageView]
 ///
-/// Some of [PhotoView] constructor options are passed direct to [PhotoViewGallery] cosntructor. Those options will affect the gallery in a whole.
+/// Some of [PhotoView] constructor options are passed direct to [PhotoViewGallery] constructor. Those options will affect the gallery in a whole.
 ///
 /// Some of the options may be defined to each image individually, such as `initialScale` or `heroAttributes`. Those must be passed via each [PhotoViewGalleryPageOptions].
 ///
@@ -47,7 +49,18 @@ typedef PhotoViewGalleryBuilder = PhotoViewGalleryPageOptions Function(
 ///       heroAttributes: const HeroAttributes(tag: "tag3"),
 ///     ),
 ///   ],
-///   loadingChild: widget.loadingChild,
+///   loadingBuilder: (context, progress) => Center(
+///            child: Container(
+///              width: 20.0,
+///              height: 20.0,
+///              child: CircularProgressIndicator(
+///                value: _progress == null
+///                    ? null
+///                    : _progress.cumulativeBytesLoaded /
+///                        _progress.expectedTotalBytes,
+///              ),
+///            ),
+///          ),
 ///   backgroundDecoration: widget.backgroundDecoration,
 ///   pageController: widget.pageController,
 ///   onPageChanged: onPageChanged,
@@ -68,7 +81,18 @@ typedef PhotoViewGalleryBuilder = PhotoViewGalleryPageOptions Function(
 ///     );
 ///   },
 ///   itemCount: galleryItems.length,
-///   loadingChild: widget.loadingChild,
+///   loadingBuilder: (context, progress) => Center(
+///            child: Container(
+///              width: 20.0,
+///              height: 20.0,
+///              child: CircularProgressIndicator(
+///                value: _progress == null
+///                    ? null
+///                    : _progress.cumulativeBytesLoaded /
+///                        _progress.expectedTotalBytes,
+///              ),
+///            ),
+///          ),
 ///   backgroundDecoration: widget.backgroundDecoration,
 ///   pageController: widget.pageController,
 ///   onPageChanged: onPageChanged,
@@ -77,9 +101,9 @@ typedef PhotoViewGalleryBuilder = PhotoViewGalleryPageOptions Function(
 class PhotoViewGallery extends StatefulWidget {
   /// Construct a gallery with static items through a list of [PhotoViewGalleryPageOptions].
   const PhotoViewGallery({
-    Key key,
-    @required this.pageOptions,
-    this.loadingChild,
+    Key? key,
+    required this.pageOptions,
+    this.loadingBuilder,
     this.backgroundDecoration,
     this.gaplessPlayback = false,
     this.reverse = false,
@@ -90,20 +114,18 @@ class PhotoViewGallery extends StatefulWidget {
     this.scrollPhysics,
     this.scrollDirection = Axis.horizontal,
     this.customSize,
-  })  : _isBuilder = false,
-        itemCount = null,
+  })  : itemCount = null,
         builder = null,
-        assert(pageOptions != null),
         super(key: key);
 
   /// Construct a gallery with dynamic items.
   ///
   /// The builder must return a [PhotoViewGalleryPageOptions].
   const PhotoViewGallery.builder({
-    Key key,
-    @required this.itemCount,
-    @required this.builder,
-    this.loadingChild,
+    Key? key,
+    required this.itemCount,
+    required this.builder,
+    this.loadingBuilder,
     this.backgroundDecoration,
     this.gaplessPlayback = false,
     this.reverse = false,
@@ -114,29 +136,28 @@ class PhotoViewGallery extends StatefulWidget {
     this.scrollPhysics,
     this.scrollDirection = Axis.horizontal,
     this.customSize,
-  })  : _isBuilder = true,
-        pageOptions = null,
+  })  : pageOptions = null,
         assert(itemCount != null),
         assert(builder != null),
         super(key: key);
 
   /// A list of options to describe the items in the gallery
-  final List<PhotoViewGalleryPageOptions> pageOptions;
+  final List<PhotoViewGalleryPageOptions>? pageOptions;
 
   /// The count of items in the gallery, only used when constructed via [PhotoViewGallery.builder]
-  final int itemCount;
+  final int? itemCount;
 
   /// Called to build items for the gallery when using [PhotoViewGallery.builder]
-  final PhotoViewGalleryBuilder builder;
+  final PhotoViewGalleryBuilder? builder;
 
   /// [ScrollPhysics] for the internal [PageView]
-  final ScrollPhysics scrollPhysics;
+  final ScrollPhysics? scrollPhysics;
 
-  /// Mirror to [PhotoView.loadingChild]
-  final Widget loadingChild;
+  /// Mirror to [PhotoView.loadingBuilder]
+  final LoadingBuilder? loadingBuilder;
 
   /// Mirror to [PhotoView.backgroundDecoration]
-  final Decoration backgroundDecoration;
+  final BoxDecoration? backgroundDecoration;
 
   /// Mirror to [PhotoView.gaplessPlayback]
   final bool gaplessPlayback;
@@ -145,24 +166,24 @@ class PhotoViewGallery extends StatefulWidget {
   final bool reverse;
 
   /// An object that controls the [PageView] inside [PhotoViewGallery]
-  final PageController pageController;
+  final PageController? pageController;
 
   /// An callback to be called on a page change
-  final PhotoViewGalleryPageChangedCallback onPageChanged;
+  final PhotoViewGalleryPageChangedCallback? onPageChanged;
 
   /// Mirror to [PhotoView.scaleStateChangedCallback]
-  final ValueChanged<PhotoViewScaleState> scaleStateChangedCallback;
+  final ValueChanged<PhotoViewScaleState>? scaleStateChangedCallback;
 
   /// Mirror to [PhotoView.enableRotation]
   final bool enableRotation;
 
   /// Mirror to [PhotoView.customSize]
-  final Size customSize;
+  final Size? customSize;
 
   /// The axis along which the [PageView] scrolls. Mirror to [PageView.scrollDirection]
   final Axis scrollDirection;
 
-  final bool _isBuilder;
+  bool get _isBuilder => builder != null;
 
   @override
   State<StatefulWidget> createState() {
@@ -171,29 +192,24 @@ class PhotoViewGallery extends StatefulWidget {
 }
 
 class _PhotoViewGalleryState extends State<PhotoViewGallery> {
-  PageController _controller;
-
-  @override
-  void initState() {
-    _controller = widget.pageController ?? PageController();
-    super.initState();
-  }
+  late final PageController _controller =
+      widget.pageController ?? PageController();
 
   void scaleStateChangedCallback(PhotoViewScaleState scaleState) {
     if (widget.scaleStateChangedCallback != null) {
-      widget.scaleStateChangedCallback(scaleState);
+      widget.scaleStateChangedCallback!(scaleState);
     }
   }
 
   int get actualPage {
-    return _controller.hasClients ? _controller.page.floor() : 0;
+    return _controller.hasClients ? _controller.page!.floor() : 0;
   }
 
   int get itemCount {
     if (widget._isBuilder) {
-      return widget.itemCount;
+      return widget.itemCount!;
     }
-    return widget.pageOptions.length;
+    return widget.pageOptions!.length;
   }
 
   @override
@@ -237,11 +253,14 @@ class _PhotoViewGalleryState extends State<PhotoViewGallery> {
             onTapDown: pageOption.onTapDown,
             gestureDetectorBehavior: pageOption.gestureDetectorBehavior,
             tightMode: pageOption.tightMode,
+            filterQuality: pageOption.filterQuality,
+            basePosition: pageOption.basePosition,
+            disableGestures: pageOption.disableGestures,
           )
         : PhotoView(
             key: ObjectKey(index),
             imageProvider: pageOption.imageProvider,
-            loadingChild: widget.loadingChild,
+            loadingBuilder: widget.loadingBuilder,
             backgroundDecoration: widget.backgroundDecoration,
             controller: pageOption.controller,
             scaleStateController: pageOption.scaleStateController,
@@ -258,6 +277,10 @@ class _PhotoViewGalleryState extends State<PhotoViewGallery> {
             onTapDown: pageOption.onTapDown,
             gestureDetectorBehavior: pageOption.gestureDetectorBehavior,
             tightMode: pageOption.tightMode,
+            filterQuality: pageOption.filterQuality,
+            basePosition: pageOption.basePosition,
+            disableGestures: pageOption.disableGestures,
+            errorBuilder: pageOption.errorBuilder,
           );
 
     return ClipRect(
@@ -268,9 +291,9 @@ class _PhotoViewGalleryState extends State<PhotoViewGallery> {
   PhotoViewGalleryPageOptions _buildPageOption(
       BuildContext context, int index) {
     if (widget._isBuilder) {
-      return widget.builder(context, index);
+      return widget.builder!(context, index);
     }
-    return widget.pageOptions[index];
+    return widget.pageOptions![index];
   }
 }
 
@@ -280,8 +303,8 @@ class _PhotoViewGalleryState extends State<PhotoViewGallery> {
 ///
 class PhotoViewGalleryPageOptions {
   PhotoViewGalleryPageOptions({
-    Key key,
-    @required this.imageProvider,
+    Key? key,
+    required this.imageProvider,
     this.heroAttributes,
     this.minScale,
     this.maxScale,
@@ -294,13 +317,16 @@ class PhotoViewGalleryPageOptions {
     this.onTapDown,
     this.gestureDetectorBehavior,
     this.tightMode,
+    this.filterQuality,
+    this.disableGestures,
+    this.errorBuilder,
   })  : child = null,
         childSize = null,
         assert(imageProvider != null);
 
   PhotoViewGalleryPageOptions.customChild({
-    @required this.child,
-    @required this.childSize,
+    required this.child,
+    this.childSize,
     this.heroAttributes,
     this.minScale,
     this.maxScale,
@@ -313,15 +339,16 @@ class PhotoViewGalleryPageOptions {
     this.onTapDown,
     this.gestureDetectorBehavior,
     this.tightMode,
-  })  : imageProvider = null,
-        assert(child != null),
-        assert(childSize != null);
+    this.filterQuality,
+    this.disableGestures,
+  })  : errorBuilder = null,
+        imageProvider = null;
 
   /// Mirror to [PhotoView.imageProvider]
-  final ImageProvider imageProvider;
+  final ImageProvider? imageProvider;
 
   /// Mirror to [PhotoView.heroAttributes]
-  final PhotoViewHeroAttributes heroAttributes;
+  final PhotoViewHeroAttributes? heroAttributes;
 
   /// Mirror to [PhotoView.minScale]
   final dynamic minScale;
@@ -333,32 +360,41 @@ class PhotoViewGalleryPageOptions {
   final dynamic initialScale;
 
   /// Mirror to [PhotoView.controller]
-  final PhotoViewController controller;
+  final PhotoViewController? controller;
 
   /// Mirror to [PhotoView.scaleStateController]
-  final PhotoViewScaleStateController scaleStateController;
+  final PhotoViewScaleStateController? scaleStateController;
 
   /// Mirror to [PhotoView.basePosition]
-  final Alignment basePosition;
+  final Alignment? basePosition;
 
   /// Mirror to [PhotoView.child]
-  final Widget child;
+  final Widget? child;
 
   /// Mirror to [PhotoView.childSize]
-  final Size childSize;
+  final Size? childSize;
 
   /// Mirror to [PhotoView.scaleStateCycle]
-  final ScaleStateCycle scaleStateCycle;
+  final ScaleStateCycle? scaleStateCycle;
 
   /// Mirror to [PhotoView.onTapUp]
-  final PhotoViewImageTapUpCallback onTapUp;
+  final PhotoViewImageTapUpCallback? onTapUp;
 
   /// Mirror to [PhotoView.onTapDown]
-  final PhotoViewImageTapDownCallback onTapDown;
+  final PhotoViewImageTapDownCallback? onTapDown;
 
   /// Mirror to [PhotoView.gestureDetectorBehavior]
-  final HitTestBehavior gestureDetectorBehavior;
+  final HitTestBehavior? gestureDetectorBehavior;
 
   /// Mirror to [PhotoView.tightMode]
-  final bool tightMode;
+  final bool? tightMode;
+
+  /// Mirror to [PhotoView.disableGestures]
+  final bool? disableGestures;
+
+  /// Quality levels for image filters.
+  final FilterQuality? filterQuality;
+
+  /// Mirror to [PhotoView.errorBuilder]
+  final ImageErrorWidgetBuilder? errorBuilder;
 }
