@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 
 import 'package:photo_view/photo_view.dart'
     show
+        PhotoViewScaleState,
         PhotoViewHeroAttributes,
         PhotoViewImageTapDownCallback,
         PhotoViewImageTapUpCallback,
@@ -21,49 +22,53 @@ const _defaultDecoration = const BoxDecoration(
 /// to user gestures, updates to  the controller state and mounts the entire PhotoView Layout
 class PhotoViewCore extends StatefulWidget {
   const PhotoViewCore({
-    Key key,
-    @required this.imageProvider,
-    @required this.backgroundDecoration,
-    @required this.gaplessPlayback,
-    @required this.heroAttributes,
-    @required this.enableRotation,
-    @required this.onTapUp,
-    @required this.onTapDown,
-    @required this.gestureDetectorBehavior,
-    @required this.controller,
-    @required this.scaleBoundaries,
-    @required this.scaleStateCycle,
-    @required this.scaleStateController,
-    @required this.basePosition,
-    @required this.tightMode,
-  })  : customChild = null,
+    Key? key,
+    required this.imageProvider,
+    required this.backgroundDecoration,
+    required this.gaplessPlayback,
+    required this.heroAttributes,
+    required this.enableRotation,
+    required this.onTapUp,
+    required this.onTapDown,
+    required this.gestureDetectorBehavior,
+    required this.controller,
+    required this.scaleBoundaries,
+    required this.scaleStateCycle,
+    required this.scaleStateController,
+    required this.basePosition,
+    required this.tightMode,
+    required this.filterQuality,
+    required this.disableGestures,
+  })   : customChild = null,
         super(key: key);
 
   const PhotoViewCore.customChild({
-    Key key,
-    @required this.customChild,
-    @required this.backgroundDecoration,
-    @required this.heroAttributes,
-    @required this.enableRotation,
-    @required this.onTapUp,
-    @required this.onTapDown,
-    @required this.gestureDetectorBehavior,
-    @required this.controller,
-    @required this.scaleBoundaries,
-    @required this.scaleStateCycle,
-    @required this.scaleStateController,
-    @required this.basePosition,
-    @required this.tightMode,
-  })  : imageProvider = null,
+    Key? key,
+    required this.customChild,
+    required this.backgroundDecoration,
+    this.heroAttributes,
+    required this.enableRotation,
+    this.onTapUp,
+    this.onTapDown,
+    this.gestureDetectorBehavior,
+    required this.controller,
+    required this.scaleBoundaries,
+    required this.scaleStateCycle,
+    required this.scaleStateController,
+    required this.basePosition,
+    required this.tightMode,
+    required this.filterQuality,
+    required this.disableGestures,
+  })   : imageProvider = null,
         gaplessPlayback = false,
         super(key: key);
 
-  final Decoration backgroundDecoration;
-  final ImageProvider imageProvider;
-  final bool gaplessPlayback;
-  final PhotoViewHeroAttributes heroAttributes;
+  final Decoration? backgroundDecoration;
+  final ImageProvider? imageProvider;
+  final bool? gaplessPlayback;
+  final PhotoViewHeroAttributes? heroAttributes;
   final bool enableRotation;
-  final Widget customChild;
+  final Widget? customChild;
 
   final PhotoViewControllerBase controller;
   final PhotoViewScaleStateController scaleStateController;
@@ -71,11 +76,14 @@ class PhotoViewCore extends StatefulWidget {
   final ScaleStateCycle scaleStateCycle;
   final Alignment basePosition;
 
-  final PhotoViewImageTapUpCallback onTapUp;
-  final PhotoViewImageTapDownCallback onTapDown;
+  final PhotoViewImageTapUpCallback? onTapUp;
+  final PhotoViewImageTapDownCallback? onTapDown;
 
-  final HitTestBehavior gestureDetectorBehavior;
+  final HitTestBehavior? gestureDetectorBehavior;
   final bool tightMode;
+  final bool disableGestures;
+
+  final FilterQuality filterQuality;
 
   @override
   State<StatefulWidget> createState() {
@@ -90,33 +98,38 @@ class PhotoViewCoreState extends State<PhotoViewCore>
         TickerProviderStateMixin,
         PhotoViewControllerDelegate,
         HitCornersDetector {
-  Offset _normalizedPosition;
-  double _scaleBefore;
-  double _rotationBefore;
+  Offset? _normalizedPosition;
+  double? _scaleBefore;
+  double? _rotationBefore;
 
-  AnimationController _scaleAnimationController;
-  Animation<double> _scaleAnimation;
+  late final AnimationController _scaleAnimationController =
+      AnimationController(vsync: this)
+        ..addListener(handleScaleAnimation)
+        ..addStatusListener(onAnimationStatus);
+  Animation<double>? _scaleAnimation;
 
-  AnimationController _positionAnimationController;
-  Animation<Offset> _positionAnimation;
+  late final AnimationController _positionAnimationController =
+      AnimationController(vsync: this)..addListener(handlePositionAnimate);
+  Animation<Offset>? _positionAnimation;
 
-  AnimationController _rotationAnimationController;
-  Animation<double> _rotationAnimation;
+  late final AnimationController _rotationAnimationController =
+      AnimationController(vsync: this)..addListener(handleRotationAnimation);
+  Animation<double>? _rotationAnimation;
 
-  PhotoViewHeroAttributes get heroAttributes => widget.heroAttributes;
+  PhotoViewHeroAttributes? get heroAttributes => widget.heroAttributes;
 
-  ScaleBoundaries cachedScaleBoundaries;
+  late ScaleBoundaries cachedScaleBoundaries = widget.scaleBoundaries;
 
   void handleScaleAnimation() {
-    scale = _scaleAnimation.value;
+    scale = _scaleAnimation!.value;
   }
 
   void handlePositionAnimate() {
-    controller.position = _positionAnimation.value;
+    controller.position = _positionAnimation!.value;
   }
 
   void handleRotationAnimation() {
-    controller.rotation = _rotationAnimation.value;
+    controller.rotation = _rotationAnimation!.value;
   }
 
   void onScaleStart(ScaleStartDetails details) {
@@ -129,16 +142,18 @@ class PhotoViewCoreState extends State<PhotoViewCore>
   }
 
   void onScaleUpdate(ScaleUpdateDetails details) {
-    final double newScale = _scaleBefore * details.scale;
-    final Offset delta = details.focalPoint - _normalizedPosition;
+    final double newScale = _scaleBefore! * details.scale;
+    final Offset delta = details.focalPoint - _normalizedPosition!;
 
     updateScaleStateFromNewScale(newScale);
 
+    //
     updateMultiple(
       scale: newScale,
       position: clampPosition(position: delta * details.scale),
-      rotation: _rotationBefore + details.rotation,
-      rotationFocusPoint: details.focalPoint,
+      rotation:
+          widget.enableRotation ? _rotationBefore! + details.rotation : null,
+      rotationFocusPoint: widget.enableRotation ? details.focalPoint : null,
     );
   }
 
@@ -177,15 +192,13 @@ class PhotoViewCoreState extends State<PhotoViewCore>
     final double magnitude = details.velocity.pixelsPerSecond.distance;
 
     // animate velocity only if there is no scale change and a significant magnitude
-    if (_scaleBefore / _scale == 1.0 && magnitude >= 400.0) {
+    if (_scaleBefore! / _scale == 1.0 && magnitude >= 400.0) {
       final Offset direction = details.velocity.pixelsPerSecond / magnitude;
       animatePosition(
         _position,
         clampPosition(position: _position + direction * 100.0),
       );
     }
-
-    checkAndSetToInitialScaleState();
   }
 
   void onDoubleTap() {
@@ -220,23 +233,22 @@ class PhotoViewCoreState extends State<PhotoViewCore>
 
   void onAnimationStatus(AnimationStatus status) {
     if (status == AnimationStatus.completed) {
-      checkAndSetToInitialScaleState();
+      onAnimationStatusCompleted();
+    }
+  }
+
+  /// Check if scale is equal to initial after scale animation update
+  void onAnimationStatusCompleted() {
+    if (scaleStateController.scaleState != PhotoViewScaleState.initial &&
+        scale == scaleBoundaries.initialScale) {
+      scaleStateController.setInvisibly(PhotoViewScaleState.initial);
     }
   }
 
   @override
   void initState() {
     super.initState();
-    _scaleAnimationController = AnimationController(vsync: this)
-      ..addListener(handleScaleAnimation);
-    _scaleAnimationController.addStatusListener(onAnimationStatus);
-
-    _positionAnimationController = AnimationController(vsync: this)
-      ..addListener(handlePositionAnimate);
-
-    _rotationAnimationController = AnimationController(vsync: this)
-      ..addListener(handleRotationAnimation);
-    startListeners();
+    initDelegate();
     addAnimateOnScaleStateUpdate(animateOnScaleStateUpdate);
 
     cachedScaleBoundaries = widget.scaleBoundaries;
@@ -281,42 +293,56 @@ class PhotoViewCoreState extends State<PhotoViewCore>
           AsyncSnapshot<PhotoViewControllerValue> snapshot,
         ) {
           if (snapshot.hasData) {
-            final PhotoViewControllerValue value = snapshot.data;
+            final PhotoViewControllerValue value = snapshot.data!;
+            final useImageScale = widget.filterQuality != FilterQuality.none;
+
+            final computedScale = useImageScale ? 1.0 : scale;
+
             final matrix = Matrix4.identity()
               ..translate(value.position.dx, value.position.dy)
-              ..scale(scale);
-            if (widget.enableRotation) {
-              matrix..rotateZ(value.rotation);
-            }
+              ..scale(computedScale)
+              ..rotateZ(value.rotation);
 
             final Widget customChildLayout = CustomSingleChildLayout(
               delegate: _CenterWithOriginalSizeDelegate(
                 scaleBoundaries.childSize,
                 basePosition,
+                useImageScale,
               ),
               child: _buildHero(),
             );
-            return PhotoViewGestureDetector(
-              child: Container(
-                constraints: widget.tightMode
-                    ? BoxConstraints.tight(scaleBoundaries.childSize * scale)
-                    : null,
-                child: Center(
-                  child: Transform(
-                    child: customChildLayout,
-                    transform: matrix,
-                    alignment: basePosition,
-                  ),
+
+            final child = Container(
+              constraints: widget.tightMode
+                  ? BoxConstraints.tight(scaleBoundaries.childSize * scale)
+                  : null,
+              child: Center(
+                child: Transform(
+                  child: customChildLayout,
+                  transform: matrix,
+                  alignment: basePosition,
                 ),
-                decoration: widget.backgroundDecoration ?? _defaultDecoration,
               ),
+              decoration: widget.backgroundDecoration ?? _defaultDecoration,
+            );
+
+            if (widget.disableGestures) {
+              return child;
+            }
+
+            return PhotoViewGestureDetector(
+              child: child,
               onDoubleTap: nextScaleState,
               onScaleStart: onScaleStart,
               onScaleUpdate: onScaleUpdate,
               onScaleEnd: onScaleEnd,
               hitDetector: this,
-              onTapUp: widget.onTapUp == null ? null : onTapUp,
-              onTapDown: widget.onTapDown == null ? null : onTapDown,
+              onTapUp: widget.onTapUp != null
+                  ? (details) => widget.onTapUp!(context, details, value)
+                  : null,
+              onTapDown: widget.onTapDown != null
+                  ? (details) => widget.onTapDown!(context, details, value)
+                  : null,
             );
           } else {
             return Container();
@@ -327,11 +353,11 @@ class PhotoViewCoreState extends State<PhotoViewCore>
   Widget _buildHero() {
     return heroAttributes != null
         ? Hero(
-            tag: heroAttributes.tag,
-            createRectTween: heroAttributes.createRectTween,
-            flightShuttleBuilder: heroAttributes.flightShuttleBuilder,
-            placeholderBuilder: heroAttributes.placeholderBuilder,
-            transitionOnUserGestures: heroAttributes.transitionOnUserGestures,
+            tag: heroAttributes!.tag,
+            createRectTween: heroAttributes!.createRectTween,
+            flightShuttleBuilder: heroAttributes!.flightShuttleBuilder,
+            placeholderBuilder: heroAttributes!.placeholderBuilder,
+            transitionOnUserGestures: heroAttributes!.transitionOnUserGestures,
             child: _buildChild(),
           )
         : _buildChild();
@@ -339,32 +365,46 @@ class PhotoViewCoreState extends State<PhotoViewCore>
 
   Widget _buildChild() {
     return widget.hasCustomChild
-        ? widget.customChild
+        ? widget.customChild!
         : Image(
-            image: widget.imageProvider,
+            image: widget.imageProvider!,
             gaplessPlayback: widget.gaplessPlayback ?? false,
+            filterQuality: widget.filterQuality,
+            width: scaleBoundaries.childSize.width * scale,
+            fit: BoxFit.contain,
           );
   }
 }
 
 class _CenterWithOriginalSizeDelegate extends SingleChildLayoutDelegate {
-  const _CenterWithOriginalSizeDelegate(this.subjectSize, this.basePosition);
+  const _CenterWithOriginalSizeDelegate(
+    this.subjectSize,
+    this.basePosition,
+    this.useImageScale,
+  );
 
   final Size subjectSize;
   final Alignment basePosition;
+  final bool useImageScale;
 
   @override
   Offset getPositionForChild(Size size, Size childSize) {
-    final double offsetX =
-        ((size.width - subjectSize.width) / 2) * (basePosition.x + 1);
-    final double offsetY =
-        ((size.height - subjectSize.height) / 2) * (basePosition.y + 1);
+    final childWidth = useImageScale ? childSize.width : subjectSize.width;
+    final childHeight = useImageScale ? childSize.height : subjectSize.height;
+
+    final halfWidth = (size.width - childWidth) / 2;
+    final halfHeight = (size.height - childHeight) / 2;
+
+    final double offsetX = halfWidth * (basePosition.x + 1);
+    final double offsetY = halfHeight * (basePosition.y + 1);
     return Offset(offsetX, offsetY);
   }
 
   @override
   BoxConstraints getConstraintsForChild(BoxConstraints constraints) {
-    return BoxConstraints.tight(subjectSize);
+    return useImageScale
+        ? const BoxConstraints()
+        : BoxConstraints.tight(subjectSize);
   }
 
   @override
@@ -378,8 +418,10 @@ class _CenterWithOriginalSizeDelegate extends SingleChildLayoutDelegate {
       other is _CenterWithOriginalSizeDelegate &&
           runtimeType == other.runtimeType &&
           subjectSize == other.subjectSize &&
-          basePosition == other.basePosition;
+          basePosition == other.basePosition &&
+          useImageScale == other.useImageScale;
 
   @override
-  int get hashCode => subjectSize.hashCode ^ basePosition.hashCode;
+  int get hashCode =>
+      subjectSize.hashCode ^ basePosition.hashCode ^ useImageScale.hashCode;
 }
