@@ -4,24 +4,24 @@ import 'package:flutter/widgets.dart';
 /// [PhotoViewImageWrapper]'s [Hero].
 class PhotoViewHeroAttributes {
   const PhotoViewHeroAttributes({
-    @required this.tag,
+    required this.tag,
     this.createRectTween,
     this.flightShuttleBuilder,
     this.placeholderBuilder,
     this.transitionOnUserGestures = false,
-  }) : assert(tag != null);
+  });
 
   /// Mirror to [Hero.tag]
   final Object tag;
 
   /// Mirror to [Hero.createRectTween]
-  final CreateRectTween createRectTween;
+  final CreateRectTween? createRectTween;
 
   /// Mirror to [Hero.flightShuttleBuilder]
-  final HeroFlightShuttleBuilder flightShuttleBuilder;
+  final HeroFlightShuttleBuilder? flightShuttleBuilder;
 
   /// Mirror to [Hero.placeholderBuilder]
-  final HeroPlaceholderBuilder placeholderBuilder;
+  final HeroPlaceholderBuilder? placeholderBuilder;
 
   /// Mirror to [Hero.transitionOnUserGestures]
   final bool transitionOnUserGestures;
