@@ -1,4 +1,9 @@
-// A class that work as a enum. It overloads the operator `*` saving the double as a multiplier.
+/// A class that work as a enum. It overloads the operator `*` saving the double as a multiplier.
+///
+/// ```
+/// PhotoViewComputedScale.contained * 2
+/// ```
+///
 class PhotoViewComputedScale {
   const PhotoViewComputedScale._internal(this._value, [this.multiplier = 1.0]);
 
@@ -13,6 +18,10 @@ class PhotoViewComputedScale {
 
   PhotoViewComputedScale operator *(double multiplier) {
     return PhotoViewComputedScale._internal(_value, multiplier);
+  }
+
+  PhotoViewComputedScale operator /(double divider) {
+    return PhotoViewComputedScale._internal(_value, 1 / divider);
   }
 
   @override
