@@ -4,7 +4,7 @@ import 'dart:ui' show Size;
 import "package:photo_view/src/photo_view_computed_scale.dart";
 import 'package:photo_view/src/photo_view_scale_state.dart';
 
-/// Given a [PhotoViewScaleState], returns a scale value considering scaleBoundaries.
+/// Given a [PhotoViewScaleState], returns a scale value considering [scaleBoundaries].
 double getScaleForScaleState(
   PhotoViewScaleState scaleState,
   ScaleBoundaries scaleBoundaries,
@@ -21,8 +21,9 @@ double getScaleForScaleState(
           scaleBoundaries);
     case PhotoViewScaleState.originalSize:
       return _clampSize(1.0, scaleBoundaries);
+    // Will never be reached
     default:
-      return null;
+      return 0;
   }
 }
 
